Fix incorrect write retry in i_doomdev.c
The doomdev specification says write() should return the total size
of the commands that were processed successfuly, not the number
of commands.

However, I_DoomDevFlushBatch uses the `done` variable as a number
of commands processed, not their size. Therefore, when adding
the result of write to done, it should first divite it by the
size of a single command, to obtain the number of commands
processed by write.